complete test for utxo selection with change output generation
