+++ rust/frameworks/tw_utxo/src/transaction/transaction_fee.rs
@@ -1,3 +1,5 @@
+// TODO: Deprecate this module, move this somewhere else.
+
 use crate::{
     sighash_computer::{TxSigningArgs, UtxoToSign},
     transaction::{transaction_interface::TxInputInterface, transaction_parts::Amount},
@@ -9,9 +11,5 @@ pub trait TransactionFee: TransactionInterface
 where
     Self::Output: TxOutputInterface,
 {
-    fn base_size(&self) -> usize;
-    fn size(&self) -> usize;
-    fn weight(&self) -> usize;
-    fn vsize(&self) -> usize;
     fn fee(&self, fee_rate: Amount) -> Amount;
 }
